MVS: remove dead code
