+++ src/extensions/mvs/tree/generic/_spec/params-schema.spec.ts
@@ -4,14 +4,12 @@
  * @author Adam Midlik <midlik@gmail.com>
  */
 
-import * as iots from 'io-ts';
-
-import { fieldValidationIssues, RequiredField, literal, nullable, paramsValidationIssues, OptionalField } from '../params-schema';
+import { OptionalField, RequiredField, SimpleParamsSchema, UnionParamsSchema, bool, fieldValidationIssues, float, int, literal, nullable, paramsValidationIssues_new, str, union } from '../params-schema';
 
 
 describe('fieldValidationIssues', () => {
     it('fieldValidationIssues string', async () => {
-        const stringField = RequiredField(iots.string, 'Testing required field stringField');
+        const stringField = RequiredField(str, 'Testing required field stringField');
         expect(fieldValidationIssues(stringField, 'hello')).toBeUndefined();
         expect(fieldValidationIssues(stringField, '')).toBeUndefined();
         expect(fieldValidationIssues(stringField, 5)).toBeTruthy();
@@ -41,7 +39,7 @@ describe('fieldValidationIssues', () => {
         expect(fieldValidationIssues(numberParam, undefined)).toBeTruthy();
     });
     it('fieldValidationIssues int', async () => {
-        const numberParam = RequiredField(iots.Integer, 'Testing required field numberParam');
+        const numberParam = RequiredField(int, 'Testing required field numberParam');
         expect(fieldValidationIssues(numberParam, 1)).toBeUndefined();
         expect(fieldValidationIssues(numberParam, 0)).toBeUndefined();
         expect(fieldValidationIssues(numberParam, 0.5)).toBeTruthy();
@@ -50,7 +48,7 @@ describe('fieldValidationIssues', () => {
         expect(fieldValidationIssues(numberParam, undefined)).toBeTruthy();
     });
     it('fieldValidationIssues union', async () => {
-        const stringOrNumberParam = RequiredField(iots.union([iots.string, iots.number]), 'Testing required field stringOrNumberParam');
+        const stringOrNumberParam = RequiredField(union([str, float]), 'Testing required field stringOrNumberParam');
         expect(fieldValidationIssues(stringOrNumberParam, 1)).toBeUndefined();
         expect(fieldValidationIssues(stringOrNumberParam, 2)).toBeUndefined();
         expect(fieldValidationIssues(stringOrNumberParam, 'hello')).toBeUndefined();
@@ -60,7 +58,7 @@ describe('fieldValidationIssues', () => {
         expect(fieldValidationIssues(stringOrNumberParam, undefined)).toBeTruthy();
     });
     it('fieldValidationIssues nullable', async () => {
-        const stringOrNullParam = RequiredField(nullable(iots.string), 'Testing required field stringOrNullParam');
+        const stringOrNullParam = RequiredField(nullable(str), 'Testing required field stringOrNullParam');
         expect(fieldValidationIssues(stringOrNullParam, 'hello')).toBeUndefined();
         expect(fieldValidationIssues(stringOrNullParam, '')).toBeUndefined();
         expect(fieldValidationIssues(stringOrNullParam, null)).toBeUndefined();
@@ -70,38 +68,74 @@ describe('fieldValidationIssues', () => {
     });
 });
 
-const schema = {
-    name: OptionalField(iots.string, 'Anonymous', 'Testing optional field name'),
-    surname: RequiredField(iots.string, 'Testing optional field surname'),
-    lunch: RequiredField(iots.boolean, 'Testing optional field lunch'),
-    age: OptionalField(iots.number, 0, 'Testing optional field age'),
-};
+const schema = SimpleParamsSchema({
+    name: OptionalField(str, 'Anonymous', 'Testing optional field name'),
+    surname: RequiredField(str, 'Testing optional field surname'),
+    lunch: RequiredField(bool, 'Testing optional field lunch'),
+    age: OptionalField(int, 0, 'Testing optional field age'),
+});
 
 describe('validateParams', () => {
     it('validateParams', async () => {
-        expect(paramsValidationIssues(schema, { surname: 'Doe', lunch: true }, { noExtra: true })).toBeUndefined();
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true }, { noExtra: true })).toBeUndefined();
-        expect(paramsValidationIssues(schema, { surname: 'Doe', lunch: true, age: 29 }, { noExtra: true })).toBeUndefined();
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29 }, { noExtra: true })).toBeUndefined();
-        expect(paramsValidationIssues(schema, {}, { noExtra: true })).toBeTruthy();
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', age: 29 }, { noExtra: true })).toBeTruthy(); // missing `lunch`
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true, age: 'old' }, { noExtra: true })).toBeTruthy(); // wrong type of `age`
-        expect(paramsValidationIssues(schema, { surname: 'Doe', lunch: true, married: false }, { noExtra: true })).toBeTruthy(); // extra param `married`
-        expect(paramsValidationIssues(schema, { surname: 'Doe', lunch: true, married: false })).toBeUndefined(); // extra param `married`
+        expect(paramsValidationIssues_new(schema, { surname: 'Doe', lunch: true }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(schema, { surname: 'Doe', lunch: true, age: 29 }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29 }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(schema, {}, { noExtra: true })).toBeTruthy();
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', age: 29 }, { noExtra: true })).toBeTruthy(); // missing `lunch`
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true, age: 'old' }, { noExtra: true })).toBeTruthy(); // wrong type of `age`
+        expect(paramsValidationIssues_new(schema, { surname: 'Doe', lunch: true, married: false }, { noExtra: true })).toBeTruthy(); // extra param `married`
+        expect(paramsValidationIssues_new(schema, { surname: 'Doe', lunch: true, married: false })).toBeUndefined(); // extra param `married`
     });
 });
 
 
 describe('validateFullParams', () => {
     it('validateFullParams', async () => {
-        expect(paramsValidationIssues(schema, { surname: 'Doe', lunch: true }, { requireAll: true, noExtra: true })).toBeTruthy();
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true }, { requireAll: true, noExtra: true })).toBeTruthy();
-        expect(paramsValidationIssues(schema, { surname: 'Doe', lunch: true, age: 29 }, { requireAll: true, noExtra: true })).toBeTruthy();
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29 }, { requireAll: true, noExtra: true })).toBeUndefined();
-        expect(paramsValidationIssues(schema, {}, { requireAll: true, noExtra: true })).toBeTruthy();
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true, age: 'old' }, { requireAll: true, noExtra: true })).toBeTruthy(); // wrong type of `age`
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29, married: true }, { requireAll: true, noExtra: true })).toBeTruthy(); // extra param `married`
-        expect(paramsValidationIssues(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29, married: true }, { requireAll: true, noExtra: false })).toBeUndefined(); // extra param `married`
+        expect(paramsValidationIssues_new(schema, { surname: 'Doe', lunch: true }, { requireAll: true, noExtra: true })).toBeTruthy();
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true }, { requireAll: true, noExtra: true })).toBeTruthy();
+        expect(paramsValidationIssues_new(schema, { surname: 'Doe', lunch: true, age: 29 }, { requireAll: true, noExtra: true })).toBeTruthy();
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29 }, { requireAll: true, noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(schema, {}, { requireAll: true, noExtra: true })).toBeTruthy();
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true, age: 'old' }, { requireAll: true, noExtra: true })).toBeTruthy(); // wrong type of `age`
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29, married: true }, { requireAll: true, noExtra: true })).toBeTruthy(); // extra param `married`
+        expect(paramsValidationIssues_new(schema, { name: 'John', surname: 'Doe', lunch: true, age: 29, married: true }, { requireAll: true, noExtra: false })).toBeUndefined(); // extra param `married`
     });
 });
 
+
+const unionSchema = UnionParamsSchema('kind', {
+    person: SimpleParamsSchema({
+        name: OptionalField(str, 'Anonymous', 'Testing optional field name'),
+        surname: RequiredField(str, 'Testing optional field surname'),
+        lunch: RequiredField(bool, 'Testing optional field lunch'),
+        age: OptionalField(int, 0, 'Testing optional field age'),
+    }),
+    object: SimpleParamsSchema({
+        weight: RequiredField(float, 'Testing optional field weight'),
+        color: OptionalField(str, 'colorless', 'Testing optional field color'),
+    }),
+});
+
+describe('validateUnionParams', () => {
+    it('validateUnionParams', async () => {
+        expect(paramsValidationIssues_new(unionSchema, { surname: 'Doe', lunch: true }, { noExtra: true })).toBeTruthy(); // missing discriminator param `kind`
+
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', surname: 'Doe', lunch: true }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', name: 'John', surname: 'Doe', lunch: true }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', surname: 'Doe', lunch: true, age: 29 }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', name: 'John', surname: 'Doe', lunch: true, age: 29 }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person' }, { noExtra: true })).toBeTruthy();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', name: 'John', surname: 'Doe', age: 29 }, { noExtra: true })).toBeTruthy(); // missing `lunch`
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', name: 'John', surname: 'Doe', lunch: true, age: 'old' }, { noExtra: true })).toBeTruthy(); // wrong type of `age`
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', surname: 'Doe', lunch: true, married: false }, { noExtra: true })).toBeTruthy(); // extra param `married`
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'person', surname: 'Doe', lunch: true, married: false })).toBeUndefined(); // extra param `married`
+
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'object', weight: 42, color: 'black' }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'object', weight: 42 }, { noExtra: true })).toBeUndefined();
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'object', color: 'black' }, { noExtra: true })).toBeTruthy(); // missing param `weight`
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'object', weight: 42, name: 'John' }, { noExtra: true })).toBeTruthy(); // extra param `name`
+
+        expect(paramsValidationIssues_new(unionSchema, { kind: 'spanish_inquisition' }, { noExtra: true })).toBeTruthy(); // unexpected value for discriminator param `kind`
+    });
+});
