+++ tsconfig.commonjs.json
@@ -6,7 +6,7 @@
         "noImplicitAny": true,
         "noImplicitThis": true,
         "sourceMap": false,
-        "noUnusedLocals": false,
+        "noUnusedLocals": true,
         "strictNullChecks": true,
         "strictFunctionTypes": true,
         "module": "CommonJS",
